requirements.txt
mypy
pylint==3.0.3
coverage
python-dotenv
kaggle~=1.5.16
pre-commit
pyspark==3.1.1
